requirements.txt
torch
numpy
pillow
torchvision
matplotlib
pygame
